% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_wfsdata.R
\name{browse_wfsdata}
\alias{browse_wfsdata}
\title{Browse WFS data from URL using \code{OWS4R::WFSClient()}}
\usage{
browse_wfsdata(url, service_version)
}
\arguments{
\item{url}{URL to a service from which to retrieve data}

\item{service_version}{Version of the service; it might be necessary to check
\code{serviceVersion} manually at time of request}
}
\value{
A data.frame (\code{pretty = TRUE}) containing available feature
  types
}
\description{
\code{browse_wfsdata()}: Browse WFS data from a URL using
\code{OWS4R::WFSClient()}.
}
\seealso{
\code{load_wfsdata()}, \code{load_maps_data()}
}
