% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw.R
\name{draw}
\alias{draw}
\title{Draw maps with \code{ggplot2} functions and \code{sf} objects}
\usage{
draw(
  area_data,
  area_col,
  area_fill,
  other_areas_data = NULL,
  other_areas_col = NULL,
  other_areas_fill = NULL,
  subdivision_data = NULL,
  subdivision_colfill = NULL,
  subdivision_col_values = NULL,
  subdivision_fill_values = NULL,
  subdivision_legend_order = 2,
  font = "",
  size = 12,
  coordinates = NULL,
  coord_size = 4,
  coord_shape = 21,
  coord_colfill = coordinates$Type,
  coord_col = "darkred",
  coord_fillcol = "red",
  label_repel = FALSE,
  label_repel_x = coordinates$Longitude,
  label_repel_y = coordinates$Latitude,
  label_repel_label = coordinates$Place,
  label_size = size/ggplot2::.pt,
  coord_sf = TRUE,
  coord_sf_xlim = c(2.5, 32.5),
  coord_sf_ylim = c(57.3, 72),
  coord_sf_expand = FALSE,
  coord_legend_order = 1,
  annotation_scale = TRUE,
  as_location = "br",
  as_width_hint = 0.3,
  as_text_family = font,
  margin_t = 0,
  margin_r = 0,
  margin_b = 0,
  margin_l = 0,
  margin_unit = "mm",
  panel_background = "#E4F4FF",
  output_filename = NULL,
  output_device = ragg::agg_png,
  output_width = 5.68,
  output_height = 6.5,
  output_resolution = 1200,
  attribution = NULL,
  attribution_size_factor = 0.5,
  attribution_colour = "#636363"
)
}
\arguments{
\item{area_data}{An \code{sf} object to be drawn with \code{ggplot2}}

\item{area_col}{An optional string for the object's \code{colour} argument}

\item{area_fill}{An optional string for the object's \code{fill} argument}

\item{other_areas_data}{An optional \code{sf} object to be drawn with
\code{ggplot2}, default is \code{NULL}}

\item{other_areas_col}{An optional string for the object's \code{colour}
argument, default is \code{NULL}}

\item{other_areas_fill}{An optional string for the object's \code{fill}
argument, default is \code{NULL}}

\item{subdivision_data}{An optional \code{sf} object to be drawn with
\code{ggplot2}, default is \code{NULL}}

\item{subdivision_colfill}{An optional vector of strings for the object's
\code{colour} and \code{fill} arguments, default is \code{NULL}}

\item{subdivision_col_values}{An optional vector of strings for the object's
\code{colour} argument, default is \code{NULL}}

\item{subdivision_fill_values}{An optional vector of strings for the object's
\code{fill} argument, default is \code{NULL}}

\item{subdivision_legend_order}{An optional integer for specifying the
object's order in the legend, default is \code{2}}

\item{font}{An optional argument for specifying the font family, default is
\code{""}}

\item{size}{An optional argument for specifying the font size, default is
\code{12}}

\item{coordinates}{An optional vector with object to be drawn with
\code{ggplot2}, default is \code{NULL}}

\item{coord_size}{An optional integer for specifying the coordinate object's
size, default is \code{4}}

\item{coord_shape}{An optional integer for specifying the coordinate object's
shape, default is \code{21}}

\item{coord_colfill}{An optional argument for specifying the
\code{coordinates} column to be passed as \code{colour} argument, default
is \code{coordinates$Type}}

\item{coord_col}{An optional string for the object's \code{colour} argument,
default is \code{"darkred"}}

\item{coord_fillcol}{An optional string for the object's \code{fill}
argument, default is \code{"red"}}

\item{label_repel}{An optional logical argument for specifying whether to add
a \code{ggrepel} repel label, default is \code{FALSE}}

\item{label_repel_x}{An optional argument for specifying the
\code{coordinates} column to be passed as \code{x} argument, default is
\code{coordinates$Longitude}}

\item{label_repel_y}{An optional argument for specifying the
\code{coordinates} column to be passed as \code{y} argument, default is
`coordinates$Latitude`}

\item{label_repel_label}{An optional argument for specifying the
\code{coordinates} column to be passed as \code{label} argument, default is
\code{coordinates$Place}}

\item{label_size}{An optional argument for specifying the label size, default
is \code{size / ggplot2::.pt}. This is due to the conversion that happens
for \code{geom_text} objects}

\item{coord_sf}{An optional logical argument for specifying whether to use
\code{coord_sf}, default is \code{TRUE}}

\item{coord_sf_xlim}{An optional vector of integers for specifying the x-axis
limits, default is \code{c(2.5, 32.5)}, which works fine for Norway}

\item{coord_sf_ylim}{An optional vector of integers for specifying the y-axis
limits, default is \code{c(57.3, 72)}, which works fine for Norway}

\item{coord_sf_expand}{An optional logical argument for specifying whether to
expand the \code{coord_sf} object, default is \code{FALSE}}

\item{coord_legend_order}{An optional integer for specifying the coordinate
object's order in the legend, default is \code{1}}

\item{annotation_scale}{An optional logical argument for specifying whether
to add a \code{ggspatial} annotation scale, default is \code{TRUE}}

\item{as_location}{An optional argument for specifying the location of the
annotation scale, default is \code{"br"}}

\item{as_width_hint}{An optional argument for specifying the width hint of
the annotation scale, default is \code{0.3}}

\item{as_text_family}{An optional argument for specifying the font family of
the annotation scale, default is \code{font}}

\item{margin_t}{An optional integer for specifying the legend.box.margin top
margin, default is \code{0}}

\item{margin_r}{An optional integer for specifying the legend.box.margin
right margin, default is \code{0}}

\item{margin_b}{An optional integer for specifying the legend.box.margin
bottom margin, default is \code{0}}

\item{margin_l}{An optional integer for specifying the legend.box.margin left
margin, default is \code{0}}

\item{margin_unit}{An optional string for specifying the legend.box.margin
unit, default is \code{"mm"}}

\item{panel_background}{An optional string for specifying the panel
background colour, default is \code{"#E4F4FF"}}

\item{output_filename}{An optional string for specifying the output filename,
default is \code{NULL}. If it is not null, the function will save the plot
as a file}

\item{output_device}{An optional function for specifying the output device,
default is \code{ragg::agg_png}}

\item{output_width}{An optional integer for specifying the output width,
default is \code{5.68}, which works fine for Norway}

\item{output_height}{An optional integer for specifying the output height,
default is \code{6.50}, which works fine for Norway}

\item{output_resolution}{An optional integer for specifying the output
resolution, default is \code{1200} DPI}

\item{attribution}{An optional string for specifying any data source or
copyright information, default is \code{NULL}}

\item{attribution_size_factor}{An optional integer for specifying the size
factor to be applied to the font size of the attribution text, default is
\code{0.5}}

\item{attribution_colour}{An optional string for specifying the colour of the
attribution text, default is \code{"#636363"} (a shade of dark grey)}
}
\value{
A \code{ggplot2} object. If \code{output_filename} is not
  \code{NULL}, the function will save the plot as a file.
}
\description{
\code{draw()}: Draw maps with \code{ggplot2} functions and \code{sf} objects.
It is based on previous versions of the
\href{https://github.com/EirikTengesdal/MapDrawingScriptforR}{\code{Map
Drawing Script for R}}, offering a consolidated method of creating maps with
a uniform style.
}
\examples{
norway <- load_maps_data(countries = "Norway")

draw(area_data = norway,
     area_col = "#512888",
     area_fill = "#D4C2ED")
}
\seealso{
\code{browse_wfsdata()}, \code{get_map_data()},
  \code{harmonise_vars()}, \code{load_coordinates()},
  \code{load_maps_data()}, \code{load_wfsdata()}, \code{merge_sf()},
  \code{theme_map()}
}
