% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMap.R
\docType{package}
\name{drawMap-package}
\alias{drawMap-package}
\title{drawMap: Easily Draw China and World Maps}
\description{
Two functions are provided to draw standard China map and World map:
1) \code{\link{drawChinaMap}} and 2) \code{\link{drawWorldMap}} (coming soon).
(Demo datasets are also provided.)
}
\author{
\href{https://www.zhihu.com/people/psychbruce/}{Han-Wu-Shuang (Bruce) Bao} (personal profile on Zhihu.com)

E-mail: \email{baohws@psych.ac.cn} or \email{psychbruce@qq.com}
}
