% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_map_data.R
\name{get_map_data}
\alias{get_map_data}
\title{Retrieve geospatial data via URL service}
\usage{
get_map_data(
  url,
  service = "wfs",
  version = NULL,
  request = "GetFeature",
  typename = NULL,
  outname = NULL,
  append = FALSE
)
}
\arguments{
\item{url}{URL to a service from which to retrieve geospatial data}

\item{service}{\code{service} list attribute for \code{httr2::url_parse},
default is \code{wfs}}

\item{version}{\code{version} list attribute for \code{httr2::url_parse},
default is \code{NULL}}

\item{request}{\code{request} list attribute for \code{httr2::url_parse},
default is \code{GetFeature}}

\item{typename}{\code{typename} list attribute for \code{httr2::url_parse},
default is \code{NULL}}

\item{outname}{Name of an optional output file with \code{sf::st_write},
default is \code{NULL}}

\item{append}{Logical, should the output layer be appended to an existing
file (\code{append = TRUE}) or overwrite layer (\code{append = FALSE}),
default is \code{FALSE}}
}
\value{
An \code{sf} object
}
\description{
\code{get_map_data()}: Retrieve data via URL service as an \code{sf} object.
Optionally choose to save a local copy of the data drawing on the
\code{sf::st_write()} function.
}
\examples{
finland_municipalities <- get_map_data(
url = "https://geo.stat.fi/geoserver/tilastointialueet/wfs",
typename = "tilastointialueet:kunta1000k")
}
\seealso{
[browse_wfsdata()], [load_wfsdata()]
}
