% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harmonise_vars.R
\name{harmonise_vars}
\alias{harmonise_vars}
\title{Harmonise variables for subsequent merging of \code{sf} objects}
\usage{
harmonise_vars(
  data,
  rename_list = NULL,
  new_var = NULL,
  new_val = NULL,
  position = "after",
  mutate_col = NULL,
  mutate_suffix = NULL,
  replace_var = NULL,
  replace_val = NULL
)
}
\arguments{
\item{data}{An \code{sf} object}

\item{rename_list}{A named list of variables to rename}

\item{new_var}{A new variable to create}

\item{new_val}{A value for the new variable}

\item{position}{Position of the new variable}

\item{mutate_col}{A variable to mutate}

\item{mutate_suffix}{A suffix to add to the mutated variable}

\item{replace_var}{A variable to replace}

\item{replace_val}{A value with which to replace the variable}
}
\value{
An \code{sf} object
}
\description{
\code{harmonise_vars()}: Harmonise variables in an \code{sf} object for
subsequent merging with other \code{sf} object(-s) with compatible variable
names. This could be relevant for example when working with GADM map data and
merging with map data from other sources.
}
