% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wfsdata.R
\name{load_wfsdata}
\alias{load_wfsdata}
\title{Load WFS data}
\usage{
load_wfsdata(url = NULL, typename = NULL, wgs84 = TRUE, dissolve = TRUE)
}
\arguments{
\item{url}{URL to a service from which to retrieve geospatial data}

\item{typename}{Name of the layer to retrieve from the service}

\item{wgs84}{Logical indicating whether to transform the data to WGS 84
format, default is \code{TRUE}}

\item{dissolve}{Logical indicating whether to dissolve the data, default is
\code{TRUE}}
}
\value{
An \code{sf} object
}
\description{
\code{load_wfsdata()}: Wrapper function for \code{get_map_data()} that
retrieves WFS data from a WFSClient URL, optionally transforms it to WGS 84
format, and optionally dissolves the data.
}
\examples{
load_wfsdata(
url = "https://geo.stat.fi/geoserver/tilastointialueet/wfs",
typename = "tilastointialueet:kunta1000k")
}
\seealso{
\code{browse_wfsdata()}, \code{get_map_data()}
}
