% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sf.R
\name{merge_sf}
\alias{merge_sf}
\title{Merge \code{sf} objects from a specified folder}
\usage{
merge_sf(
  folder,
  layer = "ADM_ADM_0",
  check_layers = FALSE,
  wgs84 = TRUE,
  other_data = NULL,
  ...
)
}
\arguments{
\item{folder}{A character string specifying the folder containing the data}

\item{layer}{A character string specifying the layer to read from the
Geopackage file}

\item{check_layers}{A logical value specifying whether to return the
\code{sf::st_layers} output for each file in the folder}

\item{wgs84}{A logical value specifying whether to transform the data to WGS
84 (EPSG:4326) coördinate reference system}

\item{other_data}{An optional \code{sf} object to add to the final \code{sf}
object}

\item{...}{Additional arguments to pass to \code{sf::st_read()}}
}
\value{
An \code{sf} object
}
\description{
\code{merge_sf()}: Read all Geopackage files in a specified folder and
combine them into a single \code{sf} object. It is tailored for importing
downloaded \href{https://gadm.org/}{GADM spatial data}.
}
\examples{
europe <- merge_sf(folder = here::here("data", "raw"))
}
