% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_map.R
\name{theme_map}
\alias{theme_map}
\title{Specify the font size and family for the text in the map plot}
\usage{
theme_map(font = "", size = 12)
}
\arguments{
\item{font}{A font name, default is \code{""}.}

\item{size}{A number, default is \code{12}.}
}
\description{
\code{theme_map()}: Specify the font size and family for the text in the map
plot drawn with \code{draw()}.
}
\seealso{
[draw()]
}
