% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawMap.R
\docType{package}
\name{drawMap-package}
\alias{drawMap-package}
\title{drawMap: Easily Draw China Maps}
\description{
One function is provided to draw standard China map:
\code{\link{drawChinaMap}} (demo datasets are also provided.)
}
\author{
\href{https://psychbruce.github.io/}{Han-Wu-Shuang (Bruce) Bao}
}
